Transition to f32 (better efficiency) and ONNX format conversion script

+++ src/layer.rs
@@ -5,13 +5,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct Layer {
-    pub weights: Array2<f64>,    // length: neuron count * inputs count
-    pub biases: Array1<f64>,     // length: neuron count
+    pub weights: Array2<f32>,    // length: neuron count * inputs count
+    pub biases: Array1<f32>,     // length: neuron count
     pub activation: Activation
 }
 
 impl Layer {
-    pub fn new(weights_flat: Vec<f64>, biases: Vec<f64>, c_inputs:usize, c_outputs: usize, activation: Activation) -> Self {
+    pub fn new(weights_flat: Vec<f32>, biases: Vec<f32>, c_inputs:usize, c_outputs: usize, activation: Activation) -> Self {
         assert_eq!(weights_flat.len(), c_inputs * c_outputs);
         assert_eq!(biases.len(), c_outputs);
 
@@ -26,7 +26,7 @@ impl Layer {
      }
 
 
-    pub fn forward(&self, inputs: &ArrayView1<f64>) -> Array1<f64> {
+    pub fn forward(&self, inputs: &ArrayView1<f32>) -> Array1<f32> {
         let mut outputs = inputs.dot(&self.weights.t()) + &self.biases;
         self.activation.apply(&mut outputs);
